Create utils.ts for functions common to all browsers
